+++ common/achievements/index.ts
@@ -13,9 +13,9 @@ import PackOfWolves from './pack-of-wolves'
 import TeamStar from './team-star'
 import {Achievement} from './types'
 import Untouchable from './untouchable'
-import UseLikeAHermit from './use-like-a-hermit'
 import {Win1, Win10, Win100, Win500, Win1000} from './wins'
 import Wipeout from './wipeout'
+import CertifiedZombie from './certified-zombie'
 
 export const ACHIEVEMENTS_LIST: Array<Achievement> = [
 	AllCards,
@@ -33,7 +33,7 @@ export const ACHIEVEMENTS_LIST: Array<Achievement> = [
 	TeamStar,
 	OreSnatcher,
 	Designer,
-	UseLikeAHermit,
+	CertifiedZombie,
 	Win1,
 	Win10,
 	Win100,
